requirements/base.txt
Django>=1.11,<2.0.0
django-environ==0.4.1
pydatajson==0.4.15
pyaml
django-ordered-model==1.4.1
google-api-python-client
django-import-export==0.5.1
-e git+git://github.com/datosgobar/django-datajsonar#egg=django_datajsonar
